lint/reorganize admin files
